Modifications to rc adjust with pandas calculations - calculating new Manning's n for each hydroid with feedback - modifying hydroTable with newly calculated dischare

+++ tools/adjust_rc_with_feedback.py
@@ -3,34 +3,80 @@ import geopandas as gpd
 from geopandas.tools import sjoin
 import os
 import rasterio
+import pandas as pd
 
 temp_workspace = r''
 HAND_CRS = 'EPSG:3857'
 
-def update_rating_curve(grouped_median, huc6):
-    pass
+def update_rating_curve(fim_directory, output_csv, htable_path, huc6):
+    df_gmed = pd.read_csv(output_csv) # read csv to import as a dataframe
+    df_gmed = df_gmed[df_gmed.hydroid != 0] # remove entries that do not have a valid hydroid
+    print(df_gmed)
+    df_htable = pd.read_csv(htable_path)
+    print(df_htable.head)
+    for index, row in df_gmed.iterrows(): # loop through the user provided dataframe row by row
+        print('row...')
+        print(row['hydroid'])
+        df_htable_hydroid = df_htable[df_htable.HydroID == row.hydroid] # filter htable for entries with matching hydroid
+        print('filtered hydrotable...')
+        print(df_htable_hydroid)
+        find_src_stage = df_htable_hydroid.loc[df_htable_hydroid['stage'].sub(row.hand).abs().idxmin()] # find closest matching stage to the user provided HAND value
+        print('found stage...')
+        print(find_src_stage)
+        # copy the corresponding htable values for the matching stage->HAND lookup
+        df_gmed.loc[index,'src_stage'] = find_src_stage.stage
+        df_gmed.loc[index,'ManningN'] = find_src_stage.ManningN
+        df_gmed.loc[index,'SLOPE'] = find_src_stage.SLOPE
+        df_gmed.loc[index,'HydraulicRadius_m'] = find_src_stage['HydraulicRadius (m)']
+        df_gmed.loc[index,'WetArea_m2'] = find_src_stage['WetArea (m2)']
+        df_gmed.loc[index,'discharge_cms'] = find_src_stage.discharge_cms
 
+    print('crosswalked df...')
+     ## Calculate Q using Manning's equation
+    df_gmed.rename(columns={'ManningN':'ManningN_default','hydroid':'HydroID'}, inplace=True) # rename the previous ManningN column
+    df_gmed['ManningN'] = df_gmed['WetArea_m2']* \
+    pow(df_gmed['HydraulicRadius_m'],2.0/3)* \
+    pow(df_gmed['SLOPE'],0.5)/df_gmed['discharge_cms']
+    print(df_gmed)
+
+    output_calc_n_csv = os.path.join(fim_directory, huc6, 'calc_src_n_vals_' + huc6 + '.csv')
+    df_gmed.to_csv(output_calc_n_csv)
+
+    df_gmed = df_gmed.groupby(["HydroID"])[['ManningN','coll_time']].median() # cacluate median ManningN to handle cases with multiple hydroid entries
+    df_htable.rename(columns={'discharge_cms':'orig_discharge_cms'}, inplace=True)
+    df_htable = df_htable.reset_index()
+    #df_htable['ManningN'] = df_htable['orig_ManningN']
+    print(df_htable)
+    df_htable = df_htable.merge(df_gmed,  how='left', on='HydroID')
+    df_htable['discharge_cms'] = df_htable['WetArea (m2)']* \
+    pow(df_htable['HydraulicRadius (m)'],2.0/3)* \
+    pow(df_htable['SLOPE'],0.5)/df_htable['ManningN']
+    #df_htable['ManningN'].fillna(df_htable.orig_ManningN, inplace=True)
+    #df_htable['discharge_cms'].fillna(df_htable.orig_discharge_cms, inplace=True)
+
+    out_htable = os.path.join(fim_directory, huc6, 'hydroTable_mod.csv')
+    df_htable.to_csv(out_htable)
 
 def ingest_points_layer(points_layer, fim_directory, wbd_path):
-    
+
     # Read wbd_path and points_layer to determine which HUC6 each point is in.
-    wbd_huc8_read = gpd.read_file(wbd_path, layer='WBDHU6')    
+    wbd_huc8_read = gpd.read_file(wbd_path, layer='WBDHU6')
     points_layer_read = gpd.read_file(points_layer)
-    
+
     # Update CRS of points_layer_read.
     points_layer_read = points_layer_read.to_crs(HAND_CRS)
     wbd_huc8_read = wbd_huc8_read.to_crs(HAND_CRS)
-    
+
     # Spatial join the two layers.
     water_edge_df = sjoin(points_layer_read, wbd_huc8_read)
-        
+
     # Convert to GeoDataFrame.
     gdf = gpd.GeoDataFrame(water_edge_df)
-    
+
     # Add two columns for X and Y.
     gdf['X'] = gdf['geometry'].x
     gdf['Y'] = gdf['geometry'].y
-            
+
     # Extract information into dictionary.
     huc6_list = []
     for index, row in gdf.iterrows():
@@ -40,11 +86,11 @@ def ingest_points_layer(points_layer, fim_directory, wbd_path):
 
     # Define coords variable to be used in point raster value attribution.
     coords = [(x,y) for x, y in zip(water_edge_df.X, water_edge_df.Y)]
-        
+
     # Define paths to relevant HUC6 HAND data.
     for huc6 in huc6_list:
         print(huc6)
-        
+
         # Define paths to relevant HUC6 HAND data and get necessary metadata for point rasterization.
         hand_path = os.path.join(fim_directory, huc6, 'hand_grid_' + huc6 + '.tif')
         if not os.path.exists(hand_path):
@@ -54,23 +100,27 @@ def ingest_points_layer(points_layer, fim_directory, wbd_path):
         if not os.path.exists(catchments_path):
             print("Catchments grid for " + huc6 + " does not exist.")
             continue
-        
+        htable_path = os.path.join(fim_directory, huc6, 'hydroTable.csv')
+        if not os.path.exists(htable_path):
+            print("hydroTable for " + huc6 + " does not exist.")
+            continue
+
 #        water_edge_df = water_edge_df[water_edge_df['HUC6'] == huc6]
-        
+
         # Use point geometry to determine pixel values at catchment and HAND grids.
         hand_src = rasterio.open(hand_path)
         water_edge_df['hand'] = [h[0] for h in hand_src.sample(coords)]
         hand_src.close()
         catchments_src = rasterio.open(catchments_path)
         water_edge_df['hydroid'] = [c[0] for c in catchments_src.sample(coords)]
-                
+
         # Get median HAND value for appropriate groups.
         water_edge_median_ds = water_edge_df.groupby(["hydroid", "flow", "submitter", "coll_time", "flow_unit"])['hand'].median()
-               
+
         output_csv = os.path.join(fim_directory, huc6, 'user_supplied_n_vals_' + huc6 + '.csv')
-        
+
         water_edge_median_ds.to_csv(output_csv)
-                
+
         # 1. Loop and find the corresponding hydroids in the Hydrotable
         # 2. Grab slope, wetted area, hydraulic radius, and feature_id that correspond with the matching hydroids and HAND value for the nearest stage
         # 3. Calculate new column for new roughness using the above info
@@ -82,31 +132,31 @@ def ingest_points_layer(points_layer, fim_directory, wbd_path):
         #    4c. Add last_updated column with timestamp where values were changed, also add "submitter" column
         # 5. What do we do in catchments that match the feature_id?
         #    5a. If these catchments already have known data, then let it use those. If not, use new calculated Ns.
-        
-        update_rating_curve(water_edge_median_ds, huc6)
+
+        update_rating_curve(fim_directory, output_csv, htable_path, huc6)
 
 
 
 if __name__ == '__main__':
-    
+
     # Parse arguments.
     parser = argparse.ArgumentParser(description='Adjusts rating curve given a shapefile containing points of known water boundary.')
     parser.add_argument('-p','--points-layer',help='Path to points layer containing known water boundary locations',required=True)
     parser.add_argument('-d','--fim-directory',help='Parent directory of FIM-required datasets.',required=True)
     parser.add_argument('-w','--wbd-path', help='Path to national HUC6 layer.',required=True)
-    
+
     # Assign variables from arguments.
     args = vars(parser.parse_args())
     points_layer = args['points_layer']
     fim_directory = args['fim_directory']
     wbd_path = args['wbd_path']
-    
+
     ingest_points_layer(points_layer, fim_directory, wbd_path)
-    
+
     # Open catchment, HAND, and point grids and determine pixel values for Hydroid, HAND value, and discharge value, respectively.
-    
+
     # Open rating curve file(s).
-    
+
     # Use three values to determine the hydroid rating curve(s) to update, then update them using a variation of Manning's Equation.
-    
+
     # Ensure the JSON rating curve is updated and saved (overwitten). Consider adding attributes to document what was performed.
